requirements.txt
annotated-types==0.7.0
environs==11.0.0
gitdb==4.0.11
GitPython==3.1.41
loguru==0.7.2
marshmallow==3.22.0
packaging==24.1
prettytable==3.11.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg-pool==3.2.2
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
PyYAML==6.0.2
smmap==5.0.1
typing_extensions==4.12.2
wcwidth==0.2.13
